Add progress info for Safari downloads
Adds a progress icon and a subtitle with download progress to Safari
downloads (.download files). A .download file is really just a folder
containing a plist and the partial download, so it's easy to extract
the data from it. I would've liked to get it working with Chrome, but
it seems like it just stores everything in a single binary file, so
I'm not sure how feasible it would be to extract that data.